+++ lib/WebDriverCommandExecutor.php
@@ -16,6 +16,7 @@
 namespace Facebook\WebDriver;
 
 use Facebook\WebDriver\Remote\WebDriverCommand;
+use Facebook\WebDriver\Remote\WebDriverResponse;
 
 /**
  * Interface for all command executor.
@@ -25,7 +26,7 @@ interface WebDriverCommandExecutor
     /**
      * @param WebDriverCommand $command
      *
-     * @return mixed
+     * @return WebDriverResponse
      */
     public function execute(WebDriverCommand $command);
 }